python/src/auth/requirements.txt
Flask
dotenv
PyJWT
bcrypt
flask_sqlalchemy
uuid